world: allow actors to shoot bullets
human actors shoot bullets when space bar is pressed. creates a
bullet with an initial velocity based on a constant bullet base
velocity in the direction the actor is facing, and the actor's
current velocity (which is now calculated each update).

bullets are objects that are added to the world. on update they
move, and on draw they are drawn