world: don't let actors walk off edge of world
the start of implementing collision detection.
the world has a collider. it adds boxes around the end of the world
to the collider. each object added to the world must implement a
get_bbox() function which returns the object's bounding box shape.

on collision between shapes callbacks are dispatched to the objects.
if an actor detects it has collided with the edge of the map it moves
to where it's not colliding.