Update to love 0.9
Update libraries to newer versions.

Minor changes to conf.lua.

Replace love2d-assets-loader as a submodule with the single loader.lua
file instead. Make the minor change to this file to port it to 0.9,
since upstream hasn't yet been ported.

This also makes the require path for it much nicer: upstream's repo
layout doesn't lend itself well to being included as a submodule.